Refactored badger profile component into separate module
